tema5/lab Big Data/requirements.txt
# Fase 1 - Scraping y procesamiento local
selenium>=4.18.0
webdriver-manager>=4.0.1
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Fase 1 - Google Sheets
gspread>=6.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# Fase 1 - Machine Learning
numpy>=1.26.0
scikit-learn>=1.4.0

# Fase 2 - Servidor Flask (AWS)
flask>=3.0.0
gunicorn>=21.2.0
